Non-working changes. Pivoting to scene-based graph.
I'm in the middle of breaking the hpgl tools out into
a seperate class and integrating a worker thread to
support the UI thread. BUT, it's obvious to me now
that a class for HPGL is silly; all we need to store
is the "PD" command polygons as a graph, and then we
can rebuild the HPGL from the void and send it to the
plotter. This system should mean that there only needs
to be one copy of the data (in the scene) which can then
be processed back out to HPGL.